refactor(frontend): move settings option configuration to new options.js file
